pyproject.toml
[tool.poetry]
name = "test-assignment-tg-bot"
version = "0.1.0"
description = ""
authors = ["2rage <the2rage@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = "^21.6"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
